Launder folded i64 constants that wrap to uint32 in the Go backend
i32.wrap_i64 lowers to uint32(<i64>), and Go rejects that conversion for
a compile-time constant above u32::MAX (surfaced by the sqlite heavy e2e:
"constant 9223372036854775807 overflows uint32"). The i64c laundering
threshold was i64::MAX, which missed every value in (u32::MAX, i64::MAX].
Widen it to u32::MAX — the wrap target and a superset of the int64
signed-view overflow threshold — so any i64 constant that could reach a
narrowing or signed cast is routed through the runtime identity helper.

+++ crates/dewasm-backend-go/src/lib.rs
@@ -1433,8 +1433,12 @@ impl<'a> Gen<'a> {
                     format!("uint32({v})")
                 }
             }
+            // An i64 constant is cast both to int64 (signed views) and, via
+            // i32.wrap_i64, to uint32; either conversion rejects a compile-time
+            // constant beyond its range, so launder anything above u32::MAX
+            // (the wrap target — a superset of the int64 overflow threshold).
             Expr::I64Const(v) => {
-                if *v > i64::MAX as u64 {
+                if *v > u32::MAX as u64 {
                     format!("{}(0x{v:x})", self.rt("i64c"))
                 } else {
                     format!("uint64({v})")
